split the onboarding screen into modules
